+++ CotEditor/Sources/PreferencesWindowController.swift
@@ -9,7 +9,7 @@
  
  ------------------------------------------------------------------------------
  
- © 2014-2016 1024jp
+ © 2014-2017 1024jp
  
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -42,10 +42,10 @@ final class PreferencesWindowController: NSWindowController {
         AppearancePaneController(),
         EditPaneController(),
         FormatPaneController(),
-        FileDropPaneController(),
+        NSStoryboard(name: "FileDropPane", bundle: nil).instantiateInitialController() as! FileDropPaneController,
         KeyBindingsPaneController(),
         PrintPaneController(),
-        IntegrationPaneController(),
+        NSStoryboard(name: "IntegrationPane", bundle: nil).instantiateInitialController() as! IntegrationPaneController,
         ]