+++ CotEditor/Sources/Tokenizer.swift
@@ -35,6 +35,8 @@ protocol TokenRepresentable {
     static var all: [Self] { get }
     
     var token: String { get }
+    var description: String { get }
+    var localizedDescription: String { get }
 }
 
 
@@ -43,6 +45,12 @@ extension TokenRepresentable where Self: RawRepresentable, Self.RawValue == Stri
     var token: String {
         
         return Self.prefix + self.rawValue + Self.suffix
+    }
+    
+    
+    var localizedDescription: String {
+        
+        return NSLocalizedString(self.description, comment: "")
     }