+++ CHANGELOG.md
@@ -7,14 +7,19 @@ develop
 
 ### New Features
 
+- Improve File Drop feature:
+    - Now, you can add a file drop setting only for a specific syntax style.
+    - Add description field.
+    - Draw capsule for pre-formatted text in the insertion format setting field.
+    - Update the default file drop settings.
 - Now, key binding snippets can set cursor position.
 
 
 ### Improvements
 
-- Draw capsule for pre-formatted text in the insertion format field in the File Drop pane.
 - Improve performance of closing large size document.
 - Add “.swift” extension to file types treaded as CotEditor script.
+- And some minor improvements and fixes.
 - [non-AppStore ver.] Update Sparkle framework to version 1.17.0.